+++ examples/hosts/app-integration/external-views/src/app/view.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { IDiceRoller } from "../../dataObject";
+import { IDiceRoller } from "../dataObject";
 
 /**
  * Render an IDiceRoller into a given div as a text character, with a button to roll it.
