+++ examples/hosts/app-integration/external-views/src/container/containerCode.ts
