+++ examples/hosts/app-integration/external-views/src/container/index.ts
@@ -3,4 +3,4 @@
  * Licensed under the MIT License.
  */
 
-export * from "./containerRuntimeFactory";
+export * from "./containerCode";
