+++ examples/hosts/app-integration/external-views/src/dataObject/index.ts
@@ -3,5 +3,4 @@
  * Licensed under the MIT License.
  */
 
-export * from "./interface";
-export * from "./model";
+export * from "./dataObject";
