--- examples/hosts/app-integration/external-views/src/dataObject/interface.ts
@@ -1,26 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-import { EventEmitter } from "events";
-
-/**
- * IDiceRoller describes the public API surface for our dice roller data object.
- */
-export interface IDiceRoller extends EventEmitter {
-    /**
-     * Get the dice value as a number.
-     */
-    readonly value: number;
-
-    /**
-     * Roll the dice.  Will cause a "diceRolled" event to be emitted.
-     */
-    roll: () => void;
-
-    /**
-     * The diceRolled event will fire whenever someone rolls the device, either locally or remotely.
-     */
-    on(event: "diceRolled", listener: () => void): this;
-}
